requirements.txt
openai==0.27.0
python-dotenv==1.0.0
python_telegram_bot==20.1
pytz==2022.7.1
requests==2.28.1
telegram==0.0.1
python-telegram-bot[job-queue]
